Create "tarball" target at fpga/ Makefile
The new target creates a .tar.gz file containing the files needed to synthesize the design in Vivado.
Thus, the tarball can be copied to a different host with the matching Vivado installation to proceed.

Usage:

$ make SUB_PROJECT=<board> tarball

The <config>.tar.gz file will be in the generated-src folder.
Just copy it to the other machine, extract it and run:

$ make FPGA_SHELLS=<path-to-fpga-shells-repo> bitstream

The tarball includes verilog sources and ip-specific tcls.
The fpga-shells repository is still needed to do the synthesis.
Just clone it and set the correct path (defaults to ../fpga-shells/).